Se modifica la estrutura del login, el chat, entre otras cosas como la integración del panel administrador...

+++ Impresiones/src/java/co/com/rempe/impresiones/negocio/delegado/VisitanteDelegado.java
@@ -37,7 +37,7 @@ public class VisitanteDelegado {
             em.getTransaction().begin();
             Visitantes visitante = new Visitantes();
             VisitantesDAO dao = new VisitantesDAO(em);
-            visitante.setIpComputadora(Integer.parseInt(ipUsuario));
+            visitante.setIpComputadora(Long.parseLong(ipUsuario));
             visitante.setFecha(new Date());
             dao.crear(visitante);
             em.getTransaction().commit();
